chore: remove license token secret (#3)
I'm not sure we need to store the token in GCP, it just needs to be set
in the k8s secret.
In out environment we set in in GCP in order to retrieve it from CI, but
that is specific to our environment.
I'm wondering if we should just remove it from here, because it can be
missleading. Let's say a customer want to update their token because
it's expiring, they will update this TF module inputs and apply. They
may think that this is good but it's just a noop. They'll have to fetch
back this secret from secretmanager and inject in back in their helm
deployment.