feat: add external workers
This adds a new input `enable_external_workers` that replaces the old
`create_compute_address_for_mqtt` that was incomplete.

If this flag is enabled, we create two additional IPs and a DNS zone to
be able to resolve the same mqtt endpoint address from both inside and
outside of the cluster.

It also create a new artifact repo and enable public read access on it.
We use the public repo as the exported launcher image variable.

I moved subnetwork in network module.