feat(CU-8698jvv8d): add helm values and k8s secret outputs
Our EKS module is providing those new outputs to make the setup easier.
The entrypoint for all the instance configuration is now the terraform
module, and we act as a pass through for variables like the
encryption_rsa_key.
This commit also add Helm values and k8s secret outputs in order to
align with the EKS module for consistency and UX.