fix: remove google provider config from module
I've removed the google provider configuration block from the module so that module consumers can set this themselves.

I've left the `region` and `project` variables because they are used by other parts of the module, but removed the `labels` variable since it's only used for provider configuration.

I've also added a separate test case in an `examples` folder so that we can configure the provider correctly in the test case.